Add fail_fast runtime error policy
This adds a new fail_fast runtime error policy to go with our existing terminate and unwind policies.

As the name suggests, fail_fast is intended to tear down the process in the fastest way possible should a runtime error occur. It uses `__builtin_trap()` on GCC and Clang, the `__fastfail()` intrinsic on MSVC, and otherwise falls back to `std::abort()`. Unlike the terminate policy, it never attempts to print a message to stderr.