Add iterable concept
An iterable is an object which can perform internal iteration, accepting a predicate which indicates if the iteration should terminate early.

Iterables need to supply an `iterate(it, pred)` function, and an `element_type(it)` function whose return type indicates the element type of the sequence. (For C++-y reasons this is a function rather than a plain alias like `using element_type = int`).

Every sequence is iterable, but we can also have iterables which are not sequences. In particular, every range can safely be iterable, because iterators are never exposed and so we never need to worry about invalidation, i.e.

```cpp
auto iterate(std::ranges::input_range auto&& rng, auto Pred) -> bool
{
    for (auto&& elem : rng) {
        if (!pred(FWD(elem)) { return false; }
    }
    return true;
}
```