Fix variant emplace with Clang19/libstdc++
I'm not quite sure what's changed with Clang 19, but it's suddenly complaining about libstdc++12's `variant` using the non-active member of a union in constexpr.

It's possible there was such a bug in libstdc++12, but it seems weird that it wasn't picked up by either GCC12 or an older version of Clang...

Anyway, hopefully this will fix it.