Add AppleClang to CI tests
The version of AppleClang shipping with XCode 16 now has complete enough C++20 support to compile Flux, so we should start testing with it regularly.

Unfortunately Github are enacting a policy of only providing a single XCode major version with each MacOS image (https://github.com/actions/runner-images/issues/10703), so if we want to test with e.g. XCode 17+ in future we'll have to run them on different images. But that's tomorrow's problem...