CMake: disable bogus GCC warning
When using designated initializers, GCC will complain that a base class has a missing initializer -- even though there is no way to actually write it. It seems like this warning is fixed in GCC15, we'll use CMake to quiet it for now for older compilers, but perhaps we should use a pragma instead?