element_type for generator<T> is T&&
Previously, if the user specified a non-reference `T` in `generator<T>`, we changed the actual element type to `const T&`, the reason being that it was possible to call `read_at()` multiple times so we couldn't safely yield an rvalue (reference).

Under the new model, we can only read each value once, so it's safe to actually give the user what they ask for -- well, almost. Instead of yielding a prvalue we actually use an xvalue to save a move.