Make iterables work with range-for
STL iterators are required to be at least movable. Since iteration contexts are typically non-copyable and non-movable, we cannot store one inside a valid iterator. This means that in order to use plain iterables with STL algorithms we need to use the `as_range` adaptor, which can hold the iter_ctx and hand out movable iterators which hold a pointer back to it.

However, range-for loops have considerably weaker requirements than the standard library concepts. So to enable iterables to be directly used with range-for loops, this commit adds "pseudoiterators" which meet the bare minimum requirements that the range-for needs. These pseudoiterators won't work with standard library algorithms, and in fact we deliberately don't add extra typedefs and things just to try to emphasise that they're not real iterators.