Update flux::iota for new concepts
For an incrementable type, iota(i) gives an infinite iterable, and iota(i, j) gives an iterable that will iterate until i == j.

If the type is additionally decrementable, then iota(i, j) will additionally be reverse iterable. If the type is additionally advanceable, then iota(i, j) will also be sized.

If the type is strongly ordered (with op<=>) then it is a multipass sequence (and in future, a collection), plus bidir if decrementable and RA if advancable.

Finally, for a type which is `std::is_arithmetic_v<T>`, iota(a) is equivalent to `iota(a, numeric_limits<T>::max())` -- that is, we treat it as a bounded collection rather than an infinite iterable.