Make infinite repeat() iterable-only
This is an interesting one, because we're actually removing functionality -- infinite repeat was previously random-access, but we can't do that when we move to collections.

It does raise the question of whether impls should be able to override (certain? all?) algorithms -- because now `repeat(x).take(n)` is iterable-only, whereas we want it to work the same as `repeat(x, n)`, i.e. be random-access. Also, repeat(x).drop(n) could be a no-op, and stride() and reverse() and...